pyproject.toml
[project]
authors = [{name = "Jad Abou-Chakra", email = "29726242+jc211@users.noreply.github.com"}]
dependencies = [
  "torch",
  "warp-lang",
  "numpy",
  "matplotlib",
  "nvtx",
  "scipy"
]
name = "live_compacted_segments"
requires-python = ">=3.11,<3.12"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
live_compacted_segments = { path = ".", editable = true }

[tool.pixi.tasks]
